+++ src/network.js
@@ -11,7 +11,7 @@ export const Networks = {
 	TESTNET: "Test SDF Network ; September 2015"
 };
 
-var current;
+var current = null;
 
 export class Network {
 	/**
@@ -19,7 +19,7 @@ export class Network {
    * stellar networks.  It also provides the {@link Network.current} class method that returns the network
    * that will be used by this process for the purposes of generating signatures.
    *
-   * The test network is the default, but you can also override the default by using the `use`,
+   * You should select network your app will use before adding the first signature. You can use the `use`,
    * `usePublicNetwork` and `useTestNetwork` helper methods.
    *
 	 * Creates a new `Network` object.
@@ -28,13 +28,6 @@ export class Network {
 	 */
 	constructor(networkPassphrase) {
 		this._networkPassphrase = networkPassphrase;
-	}
-
-	/**
-	 * Use default network (right now default network is `testnet`).
-	 */
-	static useDefault() {
-		this.useTestNetwork();
 	}
 
 	/**
@@ -83,5 +76,3 @@ export class Network {
 		return hash(this.networkPassphrase());
 	}
 }
-
-Network.useDefault();