+++ test/unit/transaction_test.js
@@ -1,3 +1,5 @@
+import crypto from 'crypto';
+
 describe('Transaction', function() {
 
   it("constructs Transaction object from a TransactionEnvelope", function(done) {
@@ -27,6 +29,27 @@ describe('Transaction', function() {
     done();
   });
 
+  beforeEach(function() {
+    StellarBase.Network.useTestNetwork();
+  })
+
+  afterEach(function() {
+    StellarBase.Network.use(null);
+  })
+
+  it("does not sign when no Network selected", function() {
+    StellarBase.Network.use(null);
+    let source      = new StellarBase.Account("GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB", "0");
+    let destination = "GDJJRRMBK4IWLEPJGIE6SXD2LP7REGZODU7WDC3I2D6MR37F4XSHBKX2";
+    let asset       = StellarBase.Asset.native();
+    let amount      = "2000";
+    let signer      = StellarBase.Keypair.random();
+
+    let tx = new StellarBase.TransactionBuilder(source)
+                .addOperation(StellarBase.Operation.payment({destination, asset, amount}))
+                .build();
+    expect(() => tx.sign(signer)).to.throw(/No network selected/);
+  });
 
   it("signs correctly", function() {
     let source      = new StellarBase.Account("GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB", "0");
@@ -45,6 +68,25 @@ describe('Transaction', function() {
     let rawSig = env.signatures()[0].signature();
     let verified = signer.verify(tx.hash(), rawSig);
     expect(verified).to.equal(true);
+  });
+
+  it("signs using hash preimage", function() {
+    let source      = new StellarBase.Account("GBBM6BKZPEHWYO3E3YKREDPQXMS4VK35YLNU7NFBRI26RAN7GI5POFBB", "0");
+    let destination = "GDJJRRMBK4IWLEPJGIE6SXD2LP7REGZODU7WDC3I2D6MR37F4XSHBKX2";
+    let asset       = StellarBase.Asset.native();
+    let amount      = "2000";
+
+    let preimage = crypto.randomBytes(64);
+    let hash = crypto.createHash('sha256').update(preimage).digest();
+
+    let tx = new StellarBase.TransactionBuilder(source)
+                .addOperation(StellarBase.Operation.payment({destination, asset, amount}))
+                .build();
+    tx.signHashX(preimage);
+
+    let env = tx.toEnvelope();
+    expectBuffersToBeEqual(env.signatures()[0].signature(), preimage);
+    expectBuffersToBeEqual(env.signatures()[0].hint(), hash.slice(hash.length - 4));
   });
 
 
@@ -72,3 +114,9 @@ describe('Transaction', function() {
   });
 
 });
+
+function expectBuffersToBeEqual(left, right) {
+  let leftHex = left.toString('hex');
+  let rightHex = right.toString('hex');
+  expect(leftHex).to.eql(rightHex);
+}