+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 As this project is pre 1.0, breaking changes may happen for minor version bumps. A breaking change will get clearly notified in this log.
 
+## 0.7.0
+
+* Support for new signer types: `sha256Hash`, `preAuthTx`.
+* `StrKey` helper class with `strkey` encoding related methods.
+* Removed deprecated methods: `Keypair.isValidPublicKey` (use `StrKey`), `Keypair.isValidSecretKey` (use `StrKey`), `Keypair.fromSeed`, `Keypair.seed`, `Keypair.rawSeed`.
+* **Breaking changes**:
+  * `Network` must be explicitly selected. Previously testnet was a default network.
+  * `Operation.setOptions()` method `signer` param changed.
+  * `Keypair.fromAccountId()` renamed to `Keypair.fromPublicKey()`.
+  * `Keypair.accountId()` renamed to `Keypair.publicKey()`.
+  * Dropping support for `End-of-Life` node versions.
+
 ## 0.6.0
 
 * **Breaking change** `ed25519` package is now optional dependency.